// src/interfaces/apiChunkInterface/repoLightChunkInterface.ts
export interface RepoLightChunkInterface {
  name: string;
  path: string;
  sha: string;
  size: number;
  url: string;
  html_url: string;
  git_url: string;
  download_url: string;
  type: string;
  _links: {
    self: string;
    git: string;
    html: string;
  };
  child_arr: RepoLightChunkInterface[];
}